Refactor Stat component: Extract StatItem and StatValue into separate files
